pr-review fix 1 - creating app.ts clean out and also system routes
