+++ src/routes/index.ts
@@ -1,0 +1,2 @@
+export { default as expensesRouter } from './expenses.router';
+export { default as systemRouter } from './system.router';
